refactor(packages): consolidate package structure — phases 1–3
Dissolve 5 misplaced or single-file packages and relocate all files to
semantically correct destinations. No logic changes; all unit tests pass.

Packages dissolved:
- wizard/        → ui/screens/onboarding/ (OnboardingSection)
- interfaces/    → enums/ (ILabelProvider, IProduceType) + base/workers/ (IDispatcherProvider)
- helper/        → workers/ (WorkStateMapper, WorkerError) + data/ (LocaleHelper)
- extensions/    → workers/ (WorkerExtensions)
- ui/viewmodels/usecases/ → ui/viewmodels/ (FertilizerViewModel flattened)

Files moved to correct packages:
- dao/MaizePerformanceRepo, dao/ProduceMarketRepo → repos/
- base/BaseEntity → entities/
- utils/Converters → database/
- utils/GeocodingService, utils/WeatherService, utils/LocationHelper → network/
- utils/RecommendationBuilder → repos/
- utils/AnimationHelper deleted (zero callers in Compose-only codebase)

Import sites updated across ~40 files. docs/ and CLAUDE.md updated to
reflect current package layout and correct DB version (v5).
