+++ app/src/main/java/com/akilimo/mobile/data/LocaleHelper.kt
@@ -1,4 +1,4 @@
-package com.akilimo.mobile.helper
+package com.akilimo.mobile.data
 
 import android.content.Context
 import android.content.ContextWrapper
