+++ app/src/main/java/com/akilimo/mobile/entities/CassavaMarketPrice.kt
@@ -4,7 +4,7 @@ package com.akilimo.mobile.entities
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.akilimo.mobile.base.BaseEntity
+import com.akilimo.mobile.entities.BaseEntity
 import com.akilimo.mobile.enums.EnumCountry
