+++ app/src/main/java/com/akilimo/mobile/entities/InvestmentAmount.kt
@@ -4,7 +4,7 @@ package com.akilimo.mobile.entities
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.akilimo.mobile.base.BaseEntity
+import com.akilimo.mobile.entities.BaseEntity
 
 @Entity(tableName = "investment_amounts")
 data class InvestmentAmount(
