+++ app/src/main/java/com/akilimo/mobile/entities/StarchFactory.kt
@@ -4,7 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.akilimo.mobile.base.BaseEntity
+import com.akilimo.mobile.entities.BaseEntity
 
 @Entity(
     tableName = "starch_factories",
