+++ app/src/main/java/com/akilimo/mobile/enums/EnumCassavaProduceType.kt
@@ -1,6 +1,6 @@
 package com.akilimo.mobile.enums
 
-import com.akilimo.mobile.interfaces.IProduceType
+import com.akilimo.mobile.enums.IProduceType
 
 enum class EnumCassavaProduceType : IProduceType {
     GARI {
