+++ app/src/main/java/com/akilimo/mobile/enums/EnumOperationType.kt
@@ -3,7 +3,7 @@ package com.akilimo.mobile.enums
 import android.content.Context
 import android.os.Parcelable
 import com.akilimo.mobile.R
-import com.akilimo.mobile.interfaces.ILabelProvider
+import com.akilimo.mobile.enums.ILabelProvider
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
