+++ app/src/main/java/com/akilimo/mobile/enums/ILabelProvider.kt
@@ -1,4 +1,4 @@
-package com.akilimo.mobile.interfaces
+package com.akilimo.mobile.enums
 
 import android.content.Context
