+++ app/src/main/java/com/akilimo/mobile/enums/IProduceType.kt
@@ -1,4 +1,4 @@
-package com.akilimo.mobile.interfaces
+package com.akilimo.mobile.enums
 
 interface IProduceType {
     fun produce(): String
