+++ app/src/main/java/com/akilimo/mobile/ui/viewmodels/OnboardingViewModel.kt
@@ -22,7 +22,7 @@ import com.akilimo.mobile.repos.AkilimoUserRepo
 import com.akilimo.mobile.repos.CurrentPracticeRepo
 import com.akilimo.mobile.repos.SelectedFertilizerRepo
 import com.akilimo.mobile.repos.UserPreferencesRepo
-import com.akilimo.mobile.wizard.OnboardingSection
+import com.akilimo.mobile.ui.screens.onboarding.OnboardingSection
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.channels.Channel
