+++ app/src/main/java/com/akilimo/mobile/workers/WorkStateMapper.kt
@@ -1,4 +1,4 @@
-package com.akilimo.mobile.helper
+package com.akilimo.mobile.workers
 
 import android.content.Context
 import com.akilimo.mobile.R
