+++ app/src/main/java/com/akilimo/mobile/workers/WorkerError.kt
@@ -1,4 +1,4 @@
-package com.akilimo.mobile.helper
+package com.akilimo.mobile.workers
 
 import android.content.Context
 import android.widget.Toast
